requirements.txt
pandas
openai
python-dotenv
tqdm
pyyaml
numpy
